refactor: useActiveTabs 마이페이지 활성화 탭 로직 분리
